Feat: Hooks personalizados
