+++ src/App.jsx
@@ -1,78 +1,10 @@
-import { useEffect, useState } from "react"
 import Header from "./components/Header"
 import Guitar from "./components/Guitar"
-import { db } from './data/db'
+import { useCart } from "./hooks/useCart"
 
 function App() {
 
-    const initialCart = () => {
-        const localStorageCart = localStorage.getItem('cart')
-        return localStorageCart ? JSON.parse(localStorageCart) : [];
-    }
-
-    const [data, setData] = useState([]);
-    const [cart, setCart] = useState(initialCart);
-
-    const MAX_ITEMS = 5;
-    const MIN_ITEMS = 1;
-
-    useEffect(() => {
-        setData(db)
-    }, [])
-
-    useEffect( () => {
-        localStorage.setItem('cart', JSON.stringify(cart));
-    }, [cart]);
-
-    function addToCart(item) {
-        const itemExists = cart.findIndex(guitar => guitar.id === item.id);
-
-        if (itemExists >= 0) {
-            if(cart[itemExists].quantity >= MAX_ITEMS) return;
-
-            const updateCart = [...cart];
-            updateCart[itemExists].quantity++;
-            setCart(updateCart);
-        } else {
-            item.quantity = 1;
-            setCart([...cart, item]);
-        }
-    }
-
-    function removeFromCart(id) {
-        setCart( prevCart => prevCart.filter(guitar => guitar.id !== id) );
-    }
-
-    function increaseQuantity(id){
-        const updatedCart = cart.map( item => {
-            if(item.id === id && item.quantity < MAX_ITEMS){
-                return{
-                    ...item,
-                    quantity: item.quantity + 1
-                }
-            }
-            return item;
-        });
-
-        setCart(updatedCart);
-    }
-
-    function decreaseQuantity(id) {
-        const updatedCart = cart.map( item => {
-            if(item.id === id && item.quantity > MIN_ITEMS){
-                return{
-                    ...item,
-                    quantity: item.quantity - 1
-                }
-            }
-            return item;
-        });
-        setCart(updatedCart);
-    }
-
-    function cleanCart() {
-        setCart([]);
-    }
+    const { data, cart, addToCart, removeFromCart, increaseQuantity, decreaseQuantity, cleanCart, isEmpty, cartTotal } = useCart();
 
     return (
         <>
@@ -82,6 +14,8 @@ function App() {
                 increaseQuantity={increaseQuantity}
                 decreaseQuantity={decreaseQuantity}
                 cleanCart={cleanCart}
+                isEmpty={isEmpty}
+                cartTotal={cartTotal}
             />
 
             <main className="container-xl mt-5">
@@ -93,7 +27,6 @@ function App() {
                             <Guitar
                                 key={guitar.id}
                                 guitar={guitar}
-                                setCart={setCart}
                                 addToCart={addToCart}
                             />
                         ))
