+++ src/components/Header.jsx
@@ -1,10 +1,4 @@
-import { useMemo } from "react";
-
-function Header({ cart, removeFromCart, increaseQuantity, decreaseQuantity, cleanCart}) {
-
-    // State derivado
-    const isEmpty = useMemo( () => cart.length === 0, [ cart ])
-    const cartTotal = useMemo( () => cart.reduce( (total, item) => total + (item.quantity * item.price), 0 ), [cart])
+function Header({ cart, removeFromCart, increaseQuantity, decreaseQuantity, cleanCart, isEmpty, cartTotal}) {
 
     return (
         <header className="py-5 header">
